Uploaded sample model output
A sample output of the CDC_challenge_model_ANDY_CLEAN python notebook, with probabilities of each county falling into the different bins, as designated by the CDC.  As you can see, it's a lot of 99.9% for bin 1 (zero cases)... this is likely because the model is overfit and maybe not randomly over sampling.  Daniel's working on making the model more complicated.